+++ src/index.ts
@@ -16,6 +16,7 @@ import * as boundedLattice from './BoundedLattice'
 import * as boundedMeetSemilattice from './BoundedMeetSemilattice'
 import * as category from './Category'
 import * as chain from './Chain'
+import * as chainRec from './ChainRec'
 import * as comonad from './Comonad'
 import * as compactable from './Compactable'
 import * as console from './Console'
@@ -152,6 +153,10 @@ export {
    * @since 3.0.0
    */
   chain,
+  /**
+   * @since 3.0.0
+   */
+  chainRec,
   /**
    * @since 3.0.0
    */